Moved all the update trajectory functions to the SysTickHandler.
The SysTickHandler has the highest interrupt priority and should
encompass the trajectory updating, since it is the highest priority.
Also added some new logic for the manual flying. I used the logic from
the simulink file to create the math for the desired thrust values.
